Move remaining exceptions to the new exceptions module
* Moved the remaining exceptions that were defined in various other
  modules to the new separate exceptions module. That way all exceptions
  are defined in their own module.
* Updated all imports to reference the new path.
