+++ insights/client/apps/malware_detection/__init__.py
@@ -21,7 +21,8 @@ from insights.client.constants import InsightsConstants as constants
 from insights.client.utilities import (
     generate_machine_id, write_data_to_file, get_time
 )
-from insights.util.subproc import call, CalledProcessError
+from insights.core.exceptions import CalledProcessError
+from insights.util.subproc import call
 
 logger = logging.getLogger(__name__)
 MIN_YARA_VERSION = "4.1.0"
