+++ insights/tests/client/apps/test_malware_detection.py
@@ -16,9 +16,10 @@ except ImportError:
     from urllib.parse import quote as urlencode  # python 3
 
 from insights.client.apps.manifests import manifests, content_types
-from insights.util.subproc import call, CalledProcessError
 from insights.client.config import InsightsConfig
+from insights.core.exceptions import CalledProcessError
 from insights.tests.helpers import getenv_bool
+from insights.util.subproc import call
 
 from insights.client.apps.malware_detection import (
     DEFAULT_MALWARE_CONFIG, MalwareDetectionClient, InsightsConnection,
