+++ insights/tests/parsers/test_corosync_cmapctl.py
@@ -1,8 +1,7 @@
 import doctest
 import pytest
 
-from insights.core.exceptions import ParseException, SkipException
-from insights.core.plugins import ContentException
+from insights.core.exceptions import ContentException, ParseException, SkipException
 from insights.parsers import corosync_cmapctl
 from insights.tests import context_wrap
