+++ insights/tests/parsers/test_puppet_ca_cert_expire_date.py
@@ -1,8 +1,7 @@
 import doctest
 import pytest
 
-from insights.core.exceptions import ParseException, SkipException
-from insights.core.plugins import ContentException
+from insights.core.exceptions import ContentException, ParseException, SkipException
 from insights.parsers import puppet_ca_cert_expire_date
 from insights.tests import context_wrap
