+++ insights/tests/parsers/test_satellite_content_hosts_count.py
@@ -1,8 +1,7 @@
 import doctest
 import pytest
 
-from insights.core.exceptions import ParseException, SkipException
-from insights.core.plugins import ContentException
+from insights.core.exceptions import ContentException, ParseException, SkipException
 from insights.parsers import satellite_content_hosts_count
 from insights.tests import context_wrap
