+++ insights/tests/parsers/test_satellite_postgresql_query.py
@@ -1,8 +1,7 @@
 import doctest
 import pytest
 
-from insights.core.exceptions import ParseException, SkipException
-from insights.core.plugins import ContentException
+from insights.core.exceptions import ContentException, ParseException, SkipException
 from insights.parsers import satellite_postgresql_query
 from insights.tests import context_wrap
